fix: move pluginNameFromPackage to shared and use it in build tool
Closes #42

Third-party scoped plugins (e.g. @tresni/opentabs-plugin-airbase) failed
adapter injection because the build tool used plugin.name as the adapter
dictionary key while the MCP server derived a different canonical name from
the npm package name. The extension received the server's derived name and
looked up adapters["tresni-airbase"] while the adapter had registered under
adapters["airbase"].

Move pluginNameFromPackage() from mcp-server/loader.ts to
@opentabs-dev/shared (alongside PLUGIN_PREFIX), making it the single
shared derivation path. The build tool now derives the adapter key from
pkgJson.name via pluginNameFromPackage() for both bundleIIFE and
hashAndFreeze, and emits a warning when the plugin class name field
doesn't match. Update discovery.test.ts to import pluginNameFromPackage
from @opentabs-dev/shared directly. Official @opentabs-dev plugins are
unaffected — pluginNameFromPackage strips the scope, producing the same
bare name that plugin.name already used.
