+++ platform/shared/src/index.ts
@@ -52,6 +52,7 @@ export {
   normalizePluginName,
   PLATFORM_PACKAGES,
   PLUGIN_PREFIX,
+  pluginNameFromPackage,
   resolvePluginPackageCandidates,
   TOOLS_FILENAME,
 } from './constants.js';
